refactor: ajuste nos diretorios
