+++ src/db/data-source.ts
@@ -1,10 +1,10 @@
 import { DataSource, DataSourceOptions } from 'typeorm';
 import { SeederOptions } from 'typeorm-extension';
 import { MainSeeder } from './seeders/main-seeder';
-import { Reader } from '../reader/reader.entity';
-import { Loan } from '../loan/loan.entity';
-import { Book } from '../book/book.entity';
-import { Auth } from '../auth/auth.entity';
+import { Reader } from 'src/modules/reader/reader.entity';
+import { Loan } from '../modules/loan/loan.entity';
+import { Book } from 'src/modules/book/book.entity';
+import { Auth } from 'src/modules/auth/auth.entity';
 import * as dotenv from 'dotenv';
 
 dotenv.config();
