+++ src/db/seeders/fake-books.seed.ts
@@ -1,5 +1,5 @@
 import { faker } from '@faker-js/faker';
-import { Book } from '../../book/book.entity';
+import { Book } from 'src/modules/book/book.entity';
 import { DataSource } from 'typeorm';
 import { Seeder, SeederFactoryManager } from 'typeorm-extension';
