+++ src/modules/auth/auth.controller.ts
@@ -14,8 +14,8 @@ import { AuthService } from './auth.service';
 import { AuthDto } from './dtos/auth.dto';
 import { Auth } from './auth.entity';
 import { SignInDto } from './dtos/sign-in.dto';
-import { Roles } from './decorators/roles.decorator';
-import { RoleGuard } from './guards/role/role.guard';
+import { Roles } from '../../common/decorators/roles.decorator';
+import { RoleGuard } from '../../common/guards/role/role.guard';
 import { AuthGuard } from '@nestjs/passport';
 import { UpdateAuthDto } from './dtos/update-auth.dto';
 import { ApiBearerAuth, ApiOperation } from '@nestjs/swagger';
