+++ src/modules/auth/auth.entity.ts
@@ -1,4 +1,4 @@
-import { AuthRoleEnum } from '../enums/auth-role.enum';
+import { AuthRoleEnum } from 'src/common/enums/auth-role.enum';
 import {
   BeforeInsert,
   BeforeUpdate,
