+++ src/modules/book/book.entity.ts
