+++ src/modules/book/dtos/find-book.dto.ts
