+++ src/modules/loan/dtos/find-loan.dto.ts
