+++ src/modules/loan/loan.service.ts
@@ -8,7 +8,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Loan } from './loan.entity';
 import { Repository } from 'typeorm';
 import { ReaderService } from '../reader/reader.service';
-import { BookService } from 'src/book/book.service';
+import { BookService } from '../book/book.service';
 import { LoanDto } from './dtos/loan.dto';
 import { Observable, throwError } from 'rxjs';
 import { FindLoanDto } from './dtos/find-loan.dto';
