+++ src/modules/reader/reader.entity.ts
